buffer: validate args to compare
