+++ lib/buffer.js
@@ -641,7 +641,7 @@ function validateOffsetType(val, name) {
 Buffer.prototype.copy = function(target, targetStart, sourceStart, sourceEnd) {
   if (!isUint8Array(target)) {
     throw new ERR_INVALID_ARG_TYPE(
-      'otherBuffer', ['Buffer', 'Uint8Array'], target);
+      'target', ['Buffer', 'Uint8Array'], target);
   }
   targetStart = validateOffsetType(targetStart, 'targetStart');
   sourceStart = validateOffsetType(sourceStart, 'sourceStart');
@@ -714,33 +714,22 @@ Buffer.prototype.compare = function compare(target,
   if (arguments.length === 1)
     return _compare(this, target);
 
-  if (targetStart === undefined)
-    targetStart = 0;
-  else if (targetStart < 0)
+  targetStart = validateOffsetType(targetStart, 'targetStart');
+  targetEnd = validateOffsetType(targetEnd, 'targetEnd') || target.length;
+  sourceStart = validateOffsetType(sourceStart, 'sourceStart');
+  sourceEnd = validateOffsetType(sourceEnd, 'sourceEnd') || this.length;
+
+  if (targetStart < 0)
     throw new ERR_OUT_OF_RANGE('targetStart', '>= 0', targetStart);
-  else
-    targetStart >>>= 0;
 
-  if (targetEnd === undefined)
-    targetEnd = target.length;
-  else if (targetEnd > target.length)
+  if (targetEnd > target.length)
     throw new ERR_OUT_OF_RANGE('targetEnd', `<= ${target.length}`, targetEnd);
-  else
-    targetEnd >>>= 0;
 
-  if (sourceStart === undefined)
-    sourceStart = 0;
-  else if (sourceStart < 0)
+  if (sourceStart < 0)
     throw new ERR_OUT_OF_RANGE('sourceStart', '>= 0', sourceStart);
-  else
-    sourceStart >>>= 0;
 
-  if (sourceEnd === undefined)
-    sourceEnd = this.length;
-  else if (sourceEnd > this.length)
+  if (sourceEnd > this.length)
     throw new ERR_OUT_OF_RANGE('sourceEnd', `<= ${this.length}`, sourceEnd);
-  else
-    sourceEnd >>>= 0;
 
   if (sourceStart >= sourceEnd)
     return (targetStart >= targetEnd ? 0 : -1);
