+++ test/parallel/test-buffer-alloc.js
@@ -964,7 +964,7 @@ common.expectsError(
   {
     code: 'ERR_INVALID_ARG_TYPE',
     type: TypeError,
-    message: 'The "otherBuffer" argument must be one of type Buffer or ' +
+    message: 'The "target" argument must be one of type Buffer or ' +
     'Uint8Array. Received type undefined'
   });
